SDI-88: ♻️  Move config to server
