+++ server/middleware/populateCurrentUser.ts
@@ -2,7 +2,7 @@ import { RequestHandler } from 'express'
 import jwtDecode from 'jwt-decode'
 
 import logger from '../../log'
-import config from '../../config'
+import config from '../config'
 import { hmppsAuthMFAUser } from '../helpers/utilities'
 
 export default function populateCurrentUser(): RequestHandler {
